Unison.Share.API.Hash

Description

Hash-related types in the Share API.

Hash types

newtype HashJWT

Constructors

HashJWT

Fields

unHashJWT :: Text

hashJWTHash :: HashJWT -> Hash32

Grab the hash out of a hash JWT.

This decodes the whole JWT, then throws away the claims; use it if you really only need the hash!

data HashJWTClaims

Constructors

HashJWTClaims

Fields

hash :: Hash32

userId :: Maybe Text

data DecodedHashJWT

A decoded hash JWT that retains the original encoded JWT.

Constructors

DecodedHashJWT

Fields

claims :: HashJWTClaims

hashJWT :: HashJWT

decodeHashJWT :: HashJWT -> DecodedHashJWT

Decode a hash JWT.

decodeHashJWTClaims :: HashJWT -> HashJWTClaims

Decode the claims out of a hash JWT.

decodedHashJWTHash :: DecodedHashJWT -> Hash32

Grab the hash out of a decoded hash JWT.